requirements.txt
openai
docx
tqdm
pdf2image
azure-search-documents 